feat: Add CreationButtons component and integrate into AdminDashboard
