+++ frontend/package.json
@@ -29,7 +29,8 @@
     "react-qr-code": "^2.0.18",
     "react-range-slider-input": "^3.2.1",
     "react-router-dom": "^7.6.0",
-    "react-webcam": "^7.2.0"
+    "react-webcam": "^7.2.0",
+    "recharts": "^3.5.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
